rename billing file
